Sent to Ann for Review
Added new pictures (some needed cropping) and latest text

I’ve updated the pictures and the text (see http://zermattapartments.badgers-place.me.uk/).  Outstanding is some formatting, the linking in of the privacy note and the google maps links … But its worth you taking a look before hand.  In particular:

* Are you ok with the amount of text to the pictures?  Gornerwald in particular ‘feels’ like a lot!
* To keep the pictures in proportion I’ve had to crop some of the smaller ones.  I don’t think its obvious which ones but have a look and see if you can spot them!  When you click on them the full size ones are shown!
* BUT - I couldn’t do this with the last picture for the studio as we would have lost a lot of the info!   What is probably better is to add another picture.  Let me know what you think
* The privacy policy is at the page http://zermattapartments.badgers-place.me.uk/Zermatt_Apartments/privacy.html  Can you have a look and let me know if you are ok with that please?

The analytics service I’ll be using is something called Matamo.  It’s free and is hosted on a server of mine which is based in  in France.  It’s an industry standard so should be pretty trouble free.   It will give you the info you want